requirements.txt
asgiref==3.8.1
Django==5.1.2
django-filter==24.3
sqlparse==0.5.1
tzdata==2024.2
